feat: add AQMP notifier
