+++ src/index.ts
@@ -6,11 +6,11 @@ import { MongoClient, Db } from 'mongodb';
 import { Clog, LOGLEVEL } from '@fdebijl/clog';
 
 import * as CONFIG from './config';
-import { itemToArticle } from './util/article';
 import { ExtendedOutput, ExtendedItem } from './domain';
 import {  } from './domain/ExtendedItem';
 import { deduplicate, processFeed } from './processors';
-import { Notifier, WebhookNotifier, NullNotifier } from './notifiers';
+import { Notifier, PubSubNotifier } from './notifiers';
+import { checkAndPropagate } from './util/checkAndPropagate';
 
 const parser = new Parser({
   headers: {'User-Agent': 'OpenTitles Scraper by contact@opentitles.info'},
@@ -38,95 +38,13 @@ const init = (): Promise<MongoClient> => {
       useNewUrlParser: true,
       useUnifiedTopology: true
     }).then((client) => {
-      dbo = client.db('opentitles');
+      dbo = client.db(CONFIG.isProd ? CONFIG.MONGO_DB_PROD : CONFIG.MONGO_DB_TEST);
       clog.log(`Connected to ${CONFIG.MONGO_URL}`, LOGLEVEL.DEBUG);
-      notifier = new WebhookNotifier();
+      notifier = new PubSubNotifier();
       resolve(client);
     }).catch((err) => {
       reject(err);
     });
-  });
-}
-
-/**
- * Find an article in the database for a given organisation and ID.
- * @param {object} find Object with org and articleid to query with the DB.
- */
-const findArticle = async (find: object): Promise<Error | Article | null> => {
-  return new Promise((resolve) => {
-    dbo.collection('articles').findOne(find, function(err, res) {
-      if (err) {
-        resolve(new Error(JSON.stringify(err)));
-      }
-
-      resolve(res)
-    });
-  });
-}
-
-/**
- * Check with the database if the given article exists and update the title if we have a new one.
- */
-const checkWithDB = (item: ExtendedItem): Promise<void> => {
-  return new Promise(async (resolve) => {
-    const find = {
-      org: item.org,
-      articleID: item.artid,
-    };
-
-    const res = await findArticle(find);
-
-    if (res instanceof Error) {
-      clog.log(res, LOGLEVEL.ERROR);
-      resolve();
-    }
-
-    if (!res) {
-      // Does not exit yet in DB
-      const newArticle = await itemToArticle(item);
-      dbo.collection('articles').insertOne(newArticle);
-      clog.log(`[${item.org}:${item.artid}] Added new article to collection`, LOGLEVEL.OFF);
-      resolve();
-    }
-
-    if (!(res instanceof Error)) {
-      if (res && res.titles) {
-        if (res.titles[res.titles.length - 1].title !== item.title) {
-          // Article was already seen but we have a new title, add the latest title
-          res.titles.push({title: item.title, datetime: moment().format('MMMM Do YYYY, h:mm:ss a'), timestamp: moment.now()});
-          dbo.collection('articles').replaceOne(find, res);
-          clog.log(`[${item.org}:${item.artid}] New title added for article`, LOGLEVEL.OFF);
-          await notifier.notifyListeners(res);
-          return;
-        } else {
-          clog.log(`No new title for [${item.org}:${item.artid}]`, LOGLEVEL.OFF);
-        }
-      }
-    }
-  });
-}
-
-/**
- * Send every item to be checked by the DB.
- */
-const checkAndPropagate = async (items: ExtendedItem[]): Promise<void> => {
-  return new Promise((resolve) => {
-    let i = 0;
-    const limit = items.length;
-
-    const check = async (): Promise<void> => {
-      if (i < limit) {
-        const item = items[i];
-        checkWithDB(item);
-        i++;
-        return check();
-      } else {
-        resolve();
-        return;
-      }
-    }
-
-    check();
   });
 }
 
@@ -167,7 +85,7 @@ const retrieveArticles = (): Promise<void> => {
             }
 
             mediumfeed = await deduplicate(mediumfeed);
-            await checkAndPropagate(mediumfeed.items);
+            await checkAndPropagate(mediumfeed.items, dbo, notifier);
 
             return nextMedium();
           });
