+++ src/util/articleIsValid.ts
@@ -1,0 +1,3 @@
+export const articleIdIsValid = (articleId: string): boolean => {
+  return !!articleId.match(/[a-z0-9]+/gi);
+}
